+++ api-webapp/src/main/kotlin/org/jeudego/pairgoth/server/Event.kt
@@ -24,7 +24,8 @@ enum class Event(val slug: String) {
     ResultUpdated("enter-result"),
     ResultsCleared("clear-results"),
     TablesRenumbered("renumber"),
-    StandingsUpdated("set-criteria")
+    StandingsUpdated("set-criteria"),
+    PairingParamsUpdated("pairing-params")
     ;
 
     companion object {